refactor: import/export 방식 수정
