--- src/components/atoms/Profile/Nickname/index.tsx
@@ -1,8 +1,0 @@
-import styles from "./Nickname.module.scss";
-
-interface NicknameProps {
-  name: string;
-}
-export default function Nickname({ name }: NicknameProps) {
-  return <div className={styles.name}>{name}</div>;
-}
