+++ src/components/atoms/Profile/ProfileNickname/index.tsx
@@ -1,0 +1,8 @@
+import styles from "./ProfileNickname.module.scss";
+
+interface NicknameProps {
+  name: string;
+}
+export const ProfileNickname = ({ name }: NicknameProps) => {
+  return <div className={styles.name}>{name}</div>;
+};
