+++ src/components/atoms/Typography/ModalTypography/index.tsx
@@ -3,6 +3,6 @@ import styles from "./ModalTypography.module.scss";
 interface TypographyProps {
   children: React.ReactNode;
 }
-export default function ModalTypography({ children }: TypographyProps) {
+export const ModalTypography = ({ children }: TypographyProps) => {
   return <div className={styles.title}>{children}</div>;
-}
+};
