+++ src/components/atoms/Video/Nickname/index.tsx
@@ -3,6 +3,6 @@ import styles from "./Nickname.module.scss";
 interface NickNameProps {
   name: string;
 }
-export default function Nickname({ name }: NickNameProps) {
+export const Nickname = ({ name }: NickNameProps) => {
   return <span className={styles.nickname}>{name}</span>;
-}
+};
