+++ src/components/atoms/Video/ViewCount/index.tsx
@@ -4,11 +4,11 @@ import PlayGraySvg from "@/assets/images/PlayGray.svg";
 interface ViewCountProps {
   count: number;
 }
-export default function ViewCount({ count }: ViewCountProps) {
+export const ViewCount = ({ count }: ViewCountProps) => {
   return (
     <div className={styles.viewCountGroup}>
       <PlayGraySvg />
       <div className={styles.viewCount}>{count}</div>
     </div>
   );
-}
+};
