+++ src/components/molecules/VideoInfo/index.tsx
@@ -1,14 +1,12 @@
-import Nickname from "@/components/atoms/Video/Nickname";
+import { LikeCount, Nickname, ViewCount } from "@/components/atoms";
 import styles from "./VideoInfo.module.scss";
-import ViewCount from "@/components/atoms/Video/ViewCount";
-import LikeCount from "@/components/atoms/Video/LikeCount";
 
 interface VideoInfoProps {
   nickname: string;
   views: number;
   likes: number;
 }
-export default function VideoInfo({ nickname, views, likes }: VideoInfoProps) {
+export const VideoInfo = ({ nickname, views, likes }: VideoInfoProps) => {
   return (
     <div className={styles.videoInfoContainer}>
       <div className={styles.videoInfoTop}>
@@ -20,4 +18,4 @@ export default function VideoInfo({ nickname, views, likes }: VideoInfoProps) {
       </div>
     </div>
   );
-}
+};
