+++ src/components/molecules/index.ts
@@ -1,3 +1,6 @@
 // molecule 컴포넌트를 모아두는 디렉토리입니다.
 // molecule 컴포넌트는 atom 컴포넌트의 조합으로 이루어진 컴포넌트입니다.
 export * from "./Modal";
+export * from "./Menu/ProfileMenuButton";
+export * from "./VideoCard";
+export * from "./VideoInfo";
